fix: hide contact form until brevo is sorted out
