+++ biome.json
@@ -1,7 +1,16 @@
 {
   "$schema": "https://biomejs.dev/schemas/1.9.4/schema.json",
   "vcs": { "enabled": false, "clientKind": "git", "useIgnoreFile": false },
-  "files": { "ignoreUnknown": false, "ignore": ["dist"] },
+  "files": {
+    "ignoreUnknown": false,
+    "ignore": [
+      "**/build/**",
+      "**/coverage/**",
+      "**/dist/**",
+      "legacy/**",
+      "packages/backend/public/frontend/**"
+    ]
+  },
   "formatter": {
     "enabled": true,
     "useEditorconfig": false,
