+++ packages/frontend/package.json
@@ -25,8 +25,10 @@
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-hook-form": "^7.56.4",
+    "react-icons": "^5.7.0",
     "react-redux": "^9.2.0",
     "react-router": "^7.4.1",
+    "react-router-dom": "^7.18.1",
     "react-zoom-pan-pinch": "^3.7.0",
     "tailwind-merge": "^3.2.0",
     "tailwindcss": "^4.0.17"
